Automotive Data

This data set contains various automotive information including engine location, dimensions, weight, engine type, number of cylinders, and other specifications.

df_automotive_data

Format

A data frame with 38 rows and 26 variables:

symboling

Symboling code for the vehicle

normalized-losses

Normalized losses in the vehicle

make

Make of the vehicle

fuel-type

Type of fuel used (e.g., gas, diesel)

aspiration

Aspiration type (e.g., std, turbo)

num-of-doors

Number of doors (e.g., two, four)

body-style

Body style (e.g., sedan, hatchback, wagon)

drive-wheels

Drive wheels (e.g., fwd, rwd, 4wd)

engine-location

Location of the engine (e.g., front, rear)

wheel-base

Wheelbase of the vehicle in inches

length

Length of the vehicle in inches

width

Width of the vehicle in inches

height

Height of the vehicle in inches

curb-weight

Curb weight of the vehicle in pounds

engine-type

Type of engine (e.g., dohc, ohcv, ohc, l)

num-of-cylinders

Number of cylinders in the engine

engine-size

Size of the engine in cubic inches

fuel-system

Fuel system used (e.g., mpfi, 2bbl, mfi, 1bbl)

bore

Diameter of the cylinders in the engine

stroke

Stroke length of the engine

compression-ratio

Compression ratio of the engine

horsepower

Horsepower generated by the engine

peak-rpm

Peak RPM of the engine

city-mpg

Miles per gallon in the city

highway-mpg

Miles per gallon on the highway

price

Price of the vehicle

Source

Downloaded from Kaggle.com by the user Ramakrishnan Srinivasan. see https://www.kaggle.com/toramky/automobile-dataset