Row-bind two data frames with different column sets

Combines two data frames or matrices by rows even when they do not share the same columns. Columns present in one input but absent in the other are added and filled with NA before binding. Row names from both inputs are preserved unless they would produce duplicates, in which case default integer row names are used.

Usage

rbind_all(df1, df2)

Arguments

df1

A data frame or matrix.

df2

A data frame or matrix.

Value

A data frame containing all rows from df1 followed by all rows from df2, with the union of both column sets. Cells where a column did not exist in the original input are NA.

Examples

df1 <- generate_correlation_matrix(n = 10)
df2 <- generate_correlation_matrix(n = 10)
names(df2)[4] <- "X11"
rbind_all(df1 = df1, df2 = df2)
#>          X1      X2       X3       X4       X5       X6       X7        X8      X9      X10       X11
#> 1   0.56452  0.1337 -0.88321  0.78811  0.08820 -1.15574  0.28684 -1.729579  1.1546  0.97088        NA
#> 2  -0.04331  0.1457  1.44852  0.54143  1.10846  0.88471  1.66258 -0.004346 -1.0435  0.73044        NA
#> 3   0.80602  1.8109 -0.57635  1.89555 -0.45304 -0.03533 -0.15028 -1.042561  1.3595 -1.42678        NA
#> 4  -0.19444  0.3577  2.10461  0.50525  1.75469  1.88400  0.25822  0.189582  2.1804  1.66821        NA
#> 5  -1.39280 -0.9008  0.33858 -0.03087 -1.26997  0.46770 -0.17143  0.965569 -0.2607  0.49954        NA
#> 6  -0.20327 -1.2738  1.13324 -0.27955  1.41286 -0.55272  0.72225 -0.051230 -1.2262 -0.25224        NA
#> 7  -1.33443  1.9489  0.86608 -0.53518 -1.69429 -0.53707  1.37049 -0.479306  2.0954  1.59720        NA
#> 8   0.17812 -0.1379  0.50634 -2.28887 -1.65488  1.44033  0.54029 -0.473291  0.3614  1.85035        NA
#> 9   0.52307  1.3065  1.74972  0.24486  0.55935 -0.52110  1.03594  0.387942 -0.1267  0.21143        NA
#> 10  0.38392  0.2499  1.74978 -0.42460  0.47507  1.25096  1.40893 -0.523746  0.2032 -1.46456        NA
#> 11  0.43685  2.2730 -0.47158       NA -1.42179 -0.24102 -0.45772  0.291862 -1.2849  0.96346 -0.660308
#> 12  0.50460 -0.2397  1.11150       NA  0.01276 -0.12493  0.16746  1.397700  1.2300  1.16752  0.224313
#> 13  0.06589  0.2432 -0.65064       NA -0.84623 -1.36050  1.03331 -0.026338  1.1180 -0.04425 -0.208820
#> 14 -0.94078 -1.1991 -0.05706       NA -0.73270 -0.82804 -1.56850 -0.444534 -0.9738 -0.84797 -0.837990
#> 15  0.72712  0.5788 -1.14436       NA  0.29778 -0.23049 -0.10633  0.250845 -0.6018 -0.69025  0.325199
#> 16 -1.20080 -0.3149  0.52862       NA -0.10462 -1.54740 -0.23946 -0.005054 -1.0371 -2.22188  1.004856
#> 17 -2.23859 -0.3058  0.86820       NA -0.45597  0.76532  0.01741  0.656860 -1.2119  0.69494 -0.001207
#> 18 -0.68530  2.4522  1.19934       NA  0.26582 -0.02931 -0.13377  1.853392  1.6376  0.93889  1.569821
#> 19  0.06429  0.2539  1.08146       NA  0.47292 -0.41860  1.58476 -2.438387 -1.4387  0.26044 -0.750364
#> 20  0.96815 -0.8693  0.42183       NA  0.76890  0.89930 -0.08286  0.146167 -2.7177 -1.27376  1.139518
row.names(df1) <- 21:30
rbind_all(df1 = df1, df2 = df2)
#>          X1      X2       X3       X4       X5       X6       X7        X8      X9      X10       X11
#> 21  0.56452  0.1337 -0.88321  0.78811  0.08820 -1.15574  0.28684 -1.729579  1.1546  0.97088        NA
#> 22 -0.04331  0.1457  1.44852  0.54143  1.10846  0.88471  1.66258 -0.004346 -1.0435  0.73044        NA
#> 23  0.80602  1.8109 -0.57635  1.89555 -0.45304 -0.03533 -0.15028 -1.042561  1.3595 -1.42678        NA
#> 24 -0.19444  0.3577  2.10461  0.50525  1.75469  1.88400  0.25822  0.189582  2.1804  1.66821        NA
#> 25 -1.39280 -0.9008  0.33858 -0.03087 -1.26997  0.46770 -0.17143  0.965569 -0.2607  0.49954        NA
#> 26 -0.20327 -1.2738  1.13324 -0.27955  1.41286 -0.55272  0.72225 -0.051230 -1.2262 -0.25224        NA
#> 27 -1.33443  1.9489  0.86608 -0.53518 -1.69429 -0.53707  1.37049 -0.479306  2.0954  1.59720        NA
#> 28  0.17812 -0.1379  0.50634 -2.28887 -1.65488  1.44033  0.54029 -0.473291  0.3614  1.85035        NA
#> 29  0.52307  1.3065  1.74972  0.24486  0.55935 -0.52110  1.03594  0.387942 -0.1267  0.21143        NA
#> 30  0.38392  0.2499  1.74978 -0.42460  0.47507  1.25096  1.40893 -0.523746  0.2032 -1.46456        NA
#> 1   0.43685  2.2730 -0.47158       NA -1.42179 -0.24102 -0.45772  0.291862 -1.2849  0.96346 -0.660308
#> 2   0.50460 -0.2397  1.11150       NA  0.01276 -0.12493  0.16746  1.397700  1.2300  1.16752  0.224313
#> 3   0.06589  0.2432 -0.65064       NA -0.84623 -1.36050  1.03331 -0.026338  1.1180 -0.04425 -0.208820
#> 4  -0.94078 -1.1991 -0.05706       NA -0.73270 -0.82804 -1.56850 -0.444534 -0.9738 -0.84797 -0.837990
#> 5   0.72712  0.5788 -1.14436       NA  0.29778 -0.23049 -0.10633  0.250845 -0.6018 -0.69025  0.325199
#> 6  -1.20080 -0.3149  0.52862       NA -0.10462 -1.54740 -0.23946 -0.005054 -1.0371 -2.22188  1.004856
#> 7  -2.23859 -0.3058  0.86820       NA -0.45597  0.76532  0.01741  0.656860 -1.2119  0.69494 -0.001207
#> 8  -0.68530  2.4522  1.19934       NA  0.26582 -0.02931 -0.13377  1.853392  1.6376  0.93889  1.569821
#> 9   0.06429  0.2539  1.08146       NA  0.47292 -0.41860  1.58476 -2.438387 -1.4387  0.26044 -0.750364
#> 10  0.96815 -0.8693  0.42183       NA  0.76890  0.89930 -0.08286  0.146167 -2.7177 -1.27376  1.139518