Scale and dummy code

Scales numeric variables between 0 and 1 and creates dummy coding for character and factor variables.

Usage

recode_scale_dummy(df, categories = 10)

Arguments

df

Dataframe containing the dataset to be scaled and dummy coded.

categories

Numeric value representing the number of unique values a vector must have to perform dummy coding. Defaults to 10.

Details

This function processes a dataframe by scaling numeric variables and creating dummy codes for character and factor variables. The numeric variables are scaled between 0 and 1,while the character and factor variables are converted to dummy variables if they have fewer unique values than the specified `categories` parameter.

The function performs the following steps: 1. Identifies numeric variables in the dataframe and scales them. 2. Identifies character and factor variables and creates dummy variables if they meet the criteria. 3. Combines the scaled numeric variables and dummy variables into a single dataframe.

The output is a dataframe with scaled numeric variables and dummy-coded character/factor variables.

Examples

# Example with the 'infert' dataset
recode_scale_dummy(infert)
#>     X6.11yrs X12..yrs X0.5yrs     age parity induced case spontaneous stratum
#> 1          0        0       1 0.21739    1.0     0.5    1         1.0 0.00000
#> 2          0        0       1 0.91304    0.0     0.5    1         0.0 0.01220
#> 3          0        0       1 0.78261    1.0     1.0    1         0.0 0.02439
#> 4          0        0       1 0.56522    0.6     1.0    1         0.0 0.03659
#> 5          1        0       0 0.60870    0.4     0.5    1         0.5 0.04878
#> 6          1        0       0 0.65217    0.6     1.0    1         0.5 0.06098
#> 7          1        0       0 0.08696    0.0     0.0    1         0.0 0.07317
#> 8          1        0       0 0.47826    0.2     0.0    1         0.0 0.08537
#> 9          1        0       0 0.00000    0.0     0.0    1         0.5 0.09756
#> 10         1        0       0 0.30435    0.2     0.0    1         0.0 0.10976
#> 11         1        0       0 0.34783    0.2     0.5    1         0.0 0.12195
#> 12         1        0       0 0.69565    0.6     1.0    1         0.5 0.13415
#> 13         1        0       0 0.43478    0.0     0.5    1         0.0 0.14634
#> 14         1        0       0 0.34783    0.4     1.0    1         0.0 0.15854
#> 15         1        0       0 0.43478    0.2     0.5    1         0.5 0.17073
#> 16         1        0       0 0.26087    0.2     1.0    1         0.0 0.18293
#> 17         1        0       0 0.39130    0.8     1.0    1         0.5 0.19512
#> 18         1        0       0 0.21739    0.0     0.0    1         0.5 0.20732
#> 19         1        0       0 0.17391    0.4     1.0    1         0.5 0.21951
#> 20         1        0       0 1.00000    0.0     0.0    1         0.5 0.23171
#> 21         1        0       0 0.82609    0.0     0.0    1         0.5 0.24390
#> 22         1        0       0 0.60870    0.2     1.0    1         0.0 0.25610
#> 23         1        0       0 0.30435    0.2     0.0    1         1.0 0.26829
#> 24         1        0       0 0.65217    0.0     0.0    1         0.5 0.28049
#> 25         1        0       0 0.26087    0.2     0.5    1         0.5 0.29268
#> 26         1        0       0 0.82609    0.2     0.0    1         1.0 0.30488
#> 27         1        0       0 0.73913    0.2     0.0    1         1.0 0.31707
#> 28         1        0       0 0.56522    0.4     0.0    1         1.0 0.32927
#> 29         1        0       0 0.30435    0.6     0.5    1         1.0 0.34146
#> 30         1        0       0 0.39130    0.6     1.0    1         0.0 0.35366
#> 31         1        0       0 0.47826    0.0     0.0    1         0.5 0.36585
#> 32         1        0       0 0.56522    0.2     0.5    1         0.0 0.37805
#> 33         1        0       0 0.91304    0.0     0.5    1         0.0 0.39024
#> 34         1        0       0 0.47826    0.2     0.0    1         1.0 0.40244
#> 35         1        0       0 0.78261    0.0     0.5    1         0.0 0.41463
#> 36         1        0       0 0.60870    0.2     0.0    1         1.0 0.42683
#> 37         1        0       0 0.65217    0.0     0.0    1         0.5 0.43902
#> 38         1        0       0 0.56522    0.4     0.5    1         1.0 0.45122
#> 39         1        0       0 0.39130    0.4     0.0    1         0.0 0.46341
#> 40         1        0       0 0.30435    0.0     0.0    1         0.5 0.47561
#> 41         1        0       0 0.78261    0.4     0.0    1         1.0 0.48780
#> 42         1        0       0 0.60870    0.0     0.0    1         0.0 0.50000
#> 43         1        0       0 0.86957    0.0     0.0    1         0.0 0.51220
#> 44         1        0       0 0.69565    0.2     0.5    1         0.5 0.52439
#> 45         0        1       0 0.39130    0.0     0.0    1         0.0 0.53659
#> 46         0        1       0 0.69565    0.0     0.5    1         0.0 0.54878
#> 47         0        1       0 0.30435    0.2     0.0    1         1.0 0.56098
#> 48         0        1       0 0.26087    0.6     1.0    1         0.0 0.57317
#> 49         0        1       0 0.21739    0.2     1.0    1         0.0 0.58537
#> 50         0        1       0 0.73913    0.4     0.0    1         1.0 0.59756
#> 51         0        1       0 0.13043    0.4     0.5    1         1.0 0.60976
#> 52         0        1       0 0.65217    0.8     0.5    1         1.0 0.62195
#> 53         0        1       0 0.26087    0.4     0.5    1         0.5 0.63415
#> 54         0        1       0 0.30435    0.0     0.0    1         0.5 0.64634
#> 55         0        1       0 0.34783    0.2     0.0    1         1.0 0.65854
#> 56         0        1       0 0.65217    0.2     0.0    1         1.0 0.67073
#> 57         0        1       0 0.30435    0.2     0.5    1         0.0 0.68293
#> 58         0        1       0 0.30435    0.2     0.0    1         1.0 0.69512
#> 59         0        1       0 0.30435    0.0     0.0    1         0.5 0.70732
#> 60         0        1       0 0.26087    0.2     0.0    1         1.0 0.71951
#> 61         0        1       0 0.60870    0.2     0.0    1         1.0 0.73171
#> 62         0        1       0 0.17391    0.0     0.0    1         0.5 0.74390
#> 63         0        1       0 0.56522    0.0     0.0    1         0.5 0.75610
#> 64         0        1       0 0.43478    0.2     0.0    1         1.0 0.76829
#> 65         0        1       0 0.21739    0.2     0.5    1         0.0 0.78049
#> 66         0        1       0 0.47826    0.0     0.0    1         0.5 0.79268
#> 67         0        1       0 0.00000    0.0     0.0    1         0.5 0.80488
#> 68         0        1       0 0.30435    0.4     0.5    1         1.0 0.81707
#> 69         0        1       0 0.69565    0.4     0.0    1         1.0 0.82927
#> 70         0        1       0 0.17391    0.0     0.5    1         0.0 0.84146
#> 71         0        1       0 0.47826    0.0     0.5    1         0.0 0.85366
#> 72         0        1       0 0.17391    0.0     0.0    1         0.5 0.86585
#> 73         0        1       0 0.43478    0.0     0.0    1         0.5 0.87805
#> 74         0        1       0 0.73913    1.0     0.0    1         1.0 0.89024
#> 75         0        1       0 0.21739    0.2     0.0    1         1.0 0.90244
#> 76         0        1       0 0.43478    0.0     0.0    1         0.5 0.91463
#> 77         0        1       0 0.43478    0.2     0.0    1         0.5 0.92683
#> 78         0        1       0 0.17391    0.0     0.5    1         0.0 0.93902
#> 79         0        1       0 0.43478    0.0     0.0    1         0.5 0.95122
#> 80         0        1       0 0.56522    0.0     0.0    1         0.5 0.96341
#> 81         0        1       0 0.60870    0.2     1.0    1         0.0 0.97561
#> 82         0        1       0 0.34783    0.0     0.0    1         0.5 0.98780
#> 83         0        1       0 0.08696    0.0     0.0    1         0.5 1.00000
#> 84         0        0       1 0.21739    1.0     1.0    0         0.0 0.00000
#> 85         0        0       1 0.91304    0.0     0.0    0         0.0 0.01220
#> 86         0        0       1 0.78261    1.0     1.0    0         0.0 0.02439
#> 87         0        0       1 0.56522    0.6     0.0    0         0.5 0.03659
#> 88         1        0       0 0.60870    0.4     1.0    0         0.0 0.04878
#> 89         1        0       0 0.65217    0.6     0.5    0         0.5 0.06098
#> 90         1        0       0 0.08696    0.0     0.0    0         0.0 0.07317
#> 91         1        0       0 0.47826    0.2     1.0    0         0.0 0.08537
#> 92         1        0       0 0.00000    0.0     0.0    0         0.5 0.09756
#> 93         1        0       0 0.30435    0.2     0.0    0         0.5 0.10976
#> 94         1        0       0 0.34783    0.2     0.0    0         0.0 0.12195
#> 95         1        0       0 0.69565    0.6     0.5    0         0.5 0.13415
#> 96         1        0       0 0.43478    0.0     0.0    0         0.0 0.14634
#> 97         1        0       0 0.34783    0.4     0.0    0         0.5 0.15854
#> 98         1        0       0 0.43478    0.2     0.5    0         0.0 0.17073
#> 99         1        0       0 0.26087    0.2     0.5    0         0.0 0.18293
#> 100        1        0       0 0.39130    0.8     0.0    0         1.0 0.19512
#> 101        1        0       0 0.21739    0.0     0.0    0         0.0 0.20732
#> 102        1        0       0 0.17391    0.4     0.0    0         0.5 0.21951
#> 103        1        0       0 1.00000    0.0     0.0    0         0.0 0.23171
#> 104        1        0       0 0.82609    0.0     0.0    0         0.0 0.24390
#> 105        1        0       0 0.60870    0.2     0.0    0         0.0 0.25610
#> 106        1        0       0 0.30435    0.2     0.0    0         0.0 0.26829
#> 107        1        0       0 0.65217    0.0     0.0    0         0.0 0.28049
#> 108        1        0       0 0.26087    0.2     0.0    0         0.5 0.29268
#> 109        1        0       0 0.82609    0.2     0.0    0         0.0 0.30488
#> 110        1        0       0 0.73913    0.2     0.0    0         0.0 0.31707
#> 111        1        0       0 0.56522    0.4     0.0    0         0.0 0.32927
#> 112        1        0       0 0.30435    0.6     0.0    0         1.0 0.34146
#> 113        1        0       0 0.39130    0.6     0.5    0         0.5 0.35366
#> 114        1        0       0 0.47826    0.0     0.0    0         0.0 0.36585
#> 115        1        0       0 0.56522    0.2     0.5    0         0.0 0.37805
#> 116        1        0       0 0.91304    0.0     0.5    0         0.0 0.39024
#> 117        1        0       0 0.47826    0.2     0.0    0         0.0 0.40244
#> 118        1        0       0 0.78261    0.0     0.0    0         0.0 0.41463
#> 119        1        0       0 0.60870    0.2     0.0    0         0.0 0.42683
#> 120        1        0       0 0.65217    0.0     0.0    0         0.0 0.43902
#> 121        1        0       0 0.56522    0.4     1.0    0         0.0 0.45122
#> 122        1        0       0 0.39130    0.4     0.0    0         1.0 0.46341
#> 123        1        0       0 0.30435    0.0     0.5    0         0.0 0.47561
#> 124        1        0       0 0.78261    0.4     0.5    0         0.0 0.48780
#> 125        1        0       0 0.60870    0.0     0.0    0         0.0 0.50000
#> 126        1        0       0 0.86957    0.0     0.0    0         0.0 0.51220
#> 127        1        0       0 0.69565    0.2     0.0    0         0.0 0.52439
#> 128        0        1       0 0.39130    0.0     0.5    0         0.0 0.53659
#> 129        0        1       0 0.69565    0.0     0.0    0         0.0 0.54878
#> 130        0        1       0 0.30435    0.2     0.5    0         0.0 0.56098
#> 131        0        1       0 0.26087    0.6     1.0    0         0.5 0.57317
#> 132        0        1       0 0.21739    0.2     0.5    0         0.0 0.58537
#> 133        0        1       0 0.73913    0.4     0.5    0         0.0 0.59756
#> 134        0        1       0 0.13043    0.4     1.0    0         0.5 0.60976
#> 135        0        1       0 0.65217    0.8     0.5    0         0.5 0.62195
#> 136        0        1       0 0.26087    0.4     0.5    0         0.5 0.63415
#> 137        0        1       0 0.30435    0.0     0.5    0         0.0 0.64634
#> 138        0        1       0 0.34783    0.2     0.5    0         0.0 0.65854
#> 139        0        1       0 0.65217    0.2     0.5    0         0.0 0.67073
#> 140        0        1       0 0.30435    0.2     0.5    0         0.5 0.68293
#> 141        0        1       0 0.30435    0.2     1.0    0         0.0 0.69512
#> 142        0        1       0 0.30435    0.0     0.5    0         0.0 0.70732
#> 143        0        1       0 0.26087    0.2     0.5    0         0.0 0.71951
#> 144        0        1       0 0.60870    0.2     1.0    0         0.0 0.73171
#> 145        0        1       0 0.17391    0.0     0.5    0         0.0 0.74390
#> 146        0        1       0 0.56522    0.0     0.0    0         0.0 0.75610
#> 147        0        1       0 0.43478    0.2     0.0    0         0.0 0.76829
#> 148        0        1       0 0.21739    0.2     0.0    0         0.5 0.78049
#> 149        0        1       0 0.47826    0.0     0.0    0         0.0 0.79268
#> 150        0        1       0 0.00000    0.0     0.0    0         0.5 0.80488
#> 151        0        1       0 0.30435    0.4     1.0    0         0.0 0.81707
#> 152        0        1       0 0.69565    0.4     0.5    0         0.5 0.82927
#> 153        0        1       0 0.17391    0.0     0.0    0         0.0 0.84146
#> 154        0        1       0 0.47826    0.0     0.5    0         0.0 0.85366
#> 155        0        1       0 0.17391    0.0     0.0    0         0.0 0.86585
#> 156        0        1       0 0.43478    0.0     0.0    0         0.5 0.87805
#> 157        0        1       0 0.21739    0.2     0.0    0         1.0 0.90244
#> 158        0        1       0 0.43478    0.0     0.0    0         0.0 0.91463
#> 159        0        1       0 0.43478    0.2     1.0    0         0.0 0.92683
#> 160        0        1       0 0.17391    0.0     0.0    0         0.0 0.93902
#> 161        0        1       0 0.43478    0.0     0.0    0         0.0 0.95122
#> 162        0        1       0 0.56522    0.0     0.0    0         0.0 0.96341
#> 163        0        1       0 0.60870    0.2     0.0    0         0.0 0.97561
#> 164        0        1       0 0.34783    0.0     0.0    0         0.5 0.98780
#> 165        0        1       0 0.08696    0.0     0.0    0         0.5 1.00000
#> 166        0        0       1 0.21739    1.0     1.0    0         0.0 0.00000
#> 167        0        0       1 0.91304    0.0     0.0    0         0.0 0.01220
#> 168        0        0       1 0.78261    1.0     1.0    0         0.0 0.02439
#> 169        0        0       1 0.56522    0.6     0.0    0         1.0 0.03659
#> 170        1        0       0 0.60870    0.4     0.0    0         0.0 0.04878
#> 171        1        0       0 0.65217    0.6     0.0    0         1.0 0.06098
#> 172        1        0       0 0.08696    0.0     0.0    0         0.0 0.07317
#> 173        1        0       0 0.47826    0.2     0.0    0         0.5 0.08537
#> 174        1        0       0 0.00000    0.0     0.5    0         0.0 0.09756
#> 175        1        0       0 0.30435    0.2     0.0    0         0.5 0.10976
#> 176        1        0       0 0.34783    0.2     0.0    0         0.5 0.12195
#> 177        1        0       0 0.69565    0.6     0.0    0         0.5 0.13415
#> 178        1        0       0 0.43478    0.0     0.0    0         0.0 0.14634
#> 179        1        0       0 0.34783    0.4     0.0    0         1.0 0.15854
#> 180        1        0       0 0.43478    0.2     0.5    0         0.0 0.17073
#> 181        1        0       0 0.26087    0.2     0.0    0         0.0 0.18293
#> 182        1        0       0 0.39130    0.8     0.5    0         1.0 0.19512
#> 183        1        0       0 0.21739    0.0     0.5    0         0.0 0.20732
#> 184        1        0       0 0.17391    0.4     0.5    0         0.5 0.21951
#> 185        1        0       0 1.00000    0.0     0.5    0         0.0 0.23171
#> 186        1        0       0 0.82609    0.0     0.0    0         0.0 0.24390
#> 187        1        0       0 0.60870    0.2     0.0    0         0.0 0.25610
#> 188        1        0       0 0.30435    0.2     1.0    0         0.0 0.26829
#> 189        1        0       0 0.65217    0.0     0.0    0         0.5 0.28049
#> 190        1        0       0 0.26087    0.2     0.0    0         1.0 0.29268
#> 191        1        0       0 0.82609    0.2     0.0    0         0.0 0.30488
#> 192        1        0       0 0.73913    0.2     0.0    0         0.0 0.31707
#> 193        1        0       0 0.56522    0.4     0.0    0         0.0 0.32927
#> 194        1        0       0 0.30435    0.6     1.0    0         0.5 0.34146
#> 195        1        0       0 0.39130    0.6     0.5    0         0.5 0.35366
#> 196        1        0       0 0.47826    0.0     0.0    0         0.0 0.36585
#> 197        1        0       0 0.56522    0.2     0.0    0         0.0 0.37805
#> 198        1        0       0 0.91304    0.0     0.0    0         0.0 0.39024
#> 199        1        0       0 0.47826    0.2     1.0    0         0.0 0.40244
#> 200        1        0       0 0.78261    0.0     0.0    0         0.0 0.41463
#> 201        1        0       0 0.60870    0.2     0.0    0         0.0 0.42683
#> 202        1        0       0 0.65217    0.0     0.0    0         0.0 0.43902
#> 203        1        0       0 0.56522    0.4     1.0    0         0.0 0.45122
#> 204        1        0       0 0.39130    0.4     0.0    0         0.5 0.46341
#> 205        1        0       0 0.30435    0.0     0.0    0         0.0 0.47561
#> 206        1        0       0 0.78261    0.4     0.0    0         0.0 0.48780
#> 207        1        0       0 0.60870    0.0     0.0    0         0.0 0.50000
#> 208        1        0       0 0.86957    0.0     0.0    0         0.0 0.51220
#> 209        1        0       0 0.69565    0.2     0.0    0         0.0 0.52439
#> 210        0        1       0 0.39130    0.0     0.0    0         0.0 0.53659
#> 211        0        1       0 0.69565    0.0     0.0    0         0.5 0.54878
#> 212        0        1       0 0.30435    0.2     0.5    0         0.0 0.56098
#> 213        0        1       0 0.26087    0.6     1.0    0         0.0 0.57317
#> 214        0        1       0 0.21739    0.2     0.5    0         0.0 0.58537
#> 215        0        1       0 0.73913    0.4     0.5    0         0.0 0.59756
#> 216        0        1       0 0.13043    0.4     1.0    0         0.0 0.60976
#> 217        0        1       0 0.65217    0.8     1.0    0         0.5 0.62195
#> 218        0        1       0 0.26087    0.4     1.0    0         0.0 0.63415
#> 219        0        1       0 0.30435    0.0     0.0    0         0.5 0.64634
#> 220        0        1       0 0.34783    0.2     0.5    0         0.5 0.65854
#> 221        0        1       0 0.65217    0.2     0.0    0         0.5 0.67073
#> 222        0        1       0 0.30435    0.2     1.0    0         0.0 0.68293
#> 223        0        1       0 0.30435    0.2     0.5    0         0.0 0.69512
#> 224        0        1       0 0.30435    0.0     0.0    0         0.0 0.70732
#> 225        0        1       0 0.26087    0.2     0.5    0         0.0 0.71951
#> 226        0        1       0 0.60870    0.2     0.5    0         0.0 0.73171
#> 227        0        1       0 0.17391    0.0     0.5    0         0.0 0.74390
#> 228        0        1       0 0.56522    0.0     0.0    0         0.0 0.75610
#> 229        0        1       0 0.43478    0.2     0.5    0         0.0 0.76829
#> 230        0        1       0 0.21739    0.2     0.0    0         1.0 0.78049
#> 231        0        1       0 0.47826    0.0     0.5    0         0.0 0.79268
#> 232        0        1       0 0.00000    0.0     0.0    0         0.0 0.80488
#> 233        0        1       0 0.30435    0.4     1.0    0         0.0 0.81707
#> 234        0        1       0 0.69565    0.4     0.0    0         1.0 0.82927
#> 235        0        1       0 0.17391    0.0     0.5    0         0.0 0.84146
#> 236        0        1       0 0.47826    0.0     0.0    0         0.0 0.85366
#> 237        0        1       0 0.17391    0.0     0.5    0         0.0 0.86585
#> 238        0        1       0 0.43478    0.0     0.0    0         0.0 0.87805
#> 239        0        1       0 0.73913    1.0     0.0    0         1.0 0.89024
#> 240        0        1       0 0.21739    0.2     0.5    0         0.5 0.90244
#> 241        0        1       0 0.43478    0.0     0.5    0         0.0 0.91463
#> 242        0        1       0 0.43478    0.2     0.0    0         0.5 0.92683
#> 243        0        1       0 0.17391    0.0     0.0    0         0.5 0.93902
#> 244        0        1       0 0.43478    0.0     0.0    0         0.5 0.95122
#> 245        0        1       0 0.56522    0.0     0.0    0         0.0 0.96341
#> 246        0        1       0 0.60870    0.2     1.0    0         0.0 0.97561
#> 247        0        1       0 0.34783    0.0     0.0    0         0.5 0.98780
#> 248        0        1       0 0.08696    0.0     0.0    0         0.5 1.00000
#>     pooled.stratum
#> 1          0.03226
#> 2          0.00000
#> 3          0.04839
#> 4          0.01613
#> 5          0.50000
#> 6          0.56452
#> 7          0.08065
#> 8          0.33871
#> 9          0.06452
#> 10         0.29032
#> 11         0.30645
#> 12         0.58065
#> 13         0.12903
#> 14         0.45161
#> 15         0.32258
#> 16         0.27419
#> 17         0.59677
#> 18         0.09677
#> 19         0.43548
#> 20         0.25806
#> 21         0.20968
#> 22         0.37097
#> 23         0.29032
#> 24         0.17742
#> 25         0.27419
#> 26         0.41935
#> 27         0.40323
#> 28         0.48387
#> 29         0.53226
#> 30         0.54839
#> 31         0.14516
#> 32         0.35484
#> 33         0.24194
#> 34         0.33871
#> 35         0.19355
#> 36         0.37097
#> 37         0.17742
#> 38         0.48387
#> 39         0.46774
#> 40         0.11290
#> 41         0.51613
#> 42         0.16129
#> 43         0.22581
#> 44         0.38710
#> 45         0.69355
#> 46         0.75806
#> 47         0.80645
#> 48         0.96774
#> 49         0.77419
#> 50         0.95161
#> 51         0.88710
#> 52         0.98387
#> 53         0.90323
#> 54         0.66129
#> 55         0.82258
#> 56         0.87097
#> 57         0.80645
#> 58         0.80645
#> 59         0.66129
#> 60         0.79032
#> 61         0.85484
#> 62         0.64516
#> 63         0.74194
#> 64         0.83871
#> 65         0.77419
#> 66         0.72581
#> 67         0.61290
#> 68         0.91935
#> 69         0.93548
#> 70         0.64516
#> 71         0.72581
#> 72         0.64516
#> 73         0.70968
#> 74         1.00000
#> 75         0.77419
#> 76         0.70968
#> 77         0.83871
#> 78         0.64516
#> 79         0.70968
#> 80         0.74194
#> 81         0.85484
#> 82         0.67742
#> 83         0.62903
#> 84         0.03226
#> 85         0.00000
#> 86         0.04839
#> 87         0.01613
#> 88         0.50000
#> 89         0.56452
#> 90         0.08065
#> 91         0.33871
#> 92         0.06452
#> 93         0.29032
#> 94         0.30645
#> 95         0.58065
#> 96         0.12903
#> 97         0.45161
#> 98         0.32258
#> 99         0.27419
#> 100        0.59677
#> 101        0.09677
#> 102        0.43548
#> 103        0.25806
#> 104        0.20968
#> 105        0.37097
#> 106        0.29032
#> 107        0.17742
#> 108        0.27419
#> 109        0.41935
#> 110        0.40323
#> 111        0.48387
#> 112        0.53226
#> 113        0.54839
#> 114        0.14516
#> 115        0.35484
#> 116        0.24194
#> 117        0.33871
#> 118        0.19355
#> 119        0.37097
#> 120        0.17742
#> 121        0.48387
#> 122        0.46774
#> 123        0.11290
#> 124        0.51613
#> 125        0.16129
#> 126        0.22581
#> 127        0.38710
#> 128        0.69355
#> 129        0.75806
#> 130        0.80645
#> 131        0.96774
#> 132        0.77419
#> 133        0.95161
#> 134        0.88710
#> 135        0.98387
#> 136        0.90323
#> 137        0.66129
#> 138        0.82258
#> 139        0.87097
#> 140        0.80645
#> 141        0.80645
#> 142        0.66129
#> 143        0.79032
#> 144        0.85484
#> 145        0.64516
#> 146        0.74194
#> 147        0.83871
#> 148        0.77419
#> 149        0.72581
#> 150        0.61290
#> 151        0.91935
#> 152        0.93548
#> 153        0.64516
#> 154        0.72581
#> 155        0.64516
#> 156        0.70968
#> 157        0.77419
#> 158        0.70968
#> 159        0.83871
#> 160        0.64516
#> 161        0.70968
#> 162        0.74194
#> 163        0.85484
#> 164        0.67742
#> 165        0.62903
#> 166        0.03226
#> 167        0.00000
#> 168        0.04839
#> 169        0.01613
#> 170        0.50000
#> 171        0.56452
#> 172        0.08065
#> 173        0.33871
#> 174        0.06452
#> 175        0.29032
#> 176        0.30645
#> 177        0.58065
#> 178        0.12903
#> 179        0.45161
#> 180        0.32258
#> 181        0.27419
#> 182        0.59677
#> 183        0.09677
#> 184        0.43548
#> 185        0.25806
#> 186        0.20968
#> 187        0.37097
#> 188        0.29032
#> 189        0.17742
#> 190        0.27419
#> 191        0.41935
#> 192        0.40323
#> 193        0.48387
#> 194        0.53226
#> 195        0.54839
#> 196        0.14516
#> 197        0.35484
#> 198        0.24194
#> 199        0.33871
#> 200        0.19355
#> 201        0.37097
#> 202        0.17742
#> 203        0.48387
#> 204        0.46774
#> 205        0.11290
#> 206        0.51613
#> 207        0.16129
#> 208        0.22581
#> 209        0.38710
#> 210        0.69355
#> 211        0.75806
#> 212        0.80645
#> 213        0.96774
#> 214        0.77419
#> 215        0.95161
#> 216        0.88710
#> 217        0.98387
#> 218        0.90323
#> 219        0.66129
#> 220        0.82258
#> 221        0.87097
#> 222        0.80645
#> 223        0.80645
#> 224        0.66129
#> 225        0.79032
#> 226        0.85484
#> 227        0.64516
#> 228        0.74194
#> 229        0.83871
#> 230        0.77419
#> 231        0.72581
#> 232        0.61290
#> 233        0.91935
#> 234        0.93548
#> 235        0.64516
#> 236        0.72581
#> 237        0.64516
#> 238        0.70968
#> 239        1.00000
#> 240        0.77419
#> 241        0.70968
#> 242        0.83871
#> 243        0.64516
#> 244        0.70968
#> 245        0.74194
#> 246        0.85484
#> 247        0.67742
#> 248        0.62903

# Example with a custom dataframe
df<-data.frame(numeric_var=c(1,2,3,4,5),
               factor_var=factor(c('A','B','A','B','C')))
recode_scale_dummy(df)
#>   A B C
#> 1 1 0 0
#> 2 0 1 0
#> 3 1 0 0
#> 4 0 1 0
#> 5 0 0 1