Scale and dummy code

Scales numeric variables between 0 and 1 and creates dummy coding for character and factor variables.

Usage

recode_scale_dummy(df, categories = 10)

Arguments

df

Dataframe containing the dataset to be scaled and dummy coded.

categories

Numeric value representing the number of unique values a vector must have to perform dummy coding. Defaults to 10.

Details

This function processes a dataframe by scaling numeric variables and creating dummy codes for character and factor variables. The numeric variables are scaled between 0 and 1,while the character and factor variables are converted to dummy variables if they have fewer unique values than the specified `categories` parameter.

The function performs the following steps: 1. Identifies numeric variables in the dataframe and scales them. 2. Identifies character and factor variables and creates dummy variables if they meet the criteria. 3. Combines the scaled numeric variables and dummy variables into a single dataframe.

The output is a dataframe with scaled numeric variables and dummy-coded character/factor variables.

Examples

# Example with the 'infert' dataset
recode_scale_dummy(infert)
#>     X6.11yrs X12..yrs X0.5yrs     age parity induced case spontaneous stratum pooled.stratum
#> 1          0        0       1 0.21739    1.0     0.5    1         1.0 0.00000        0.03226
#> 2          0        0       1 0.91304    0.0     0.5    1         0.0 0.01220        0.00000
#> 3          0        0       1 0.78261    1.0     1.0    1         0.0 0.02439        0.04839
#> 4          0        0       1 0.56522    0.6     1.0    1         0.0 0.03659        0.01613
#> 5          1        0       0 0.60870    0.4     0.5    1         0.5 0.04878        0.50000
#> 6          1        0       0 0.65217    0.6     1.0    1         0.5 0.06098        0.56452
#> 7          1        0       0 0.08696    0.0     0.0    1         0.0 0.07317        0.08065
#> 8          1        0       0 0.47826    0.2     0.0    1         0.0 0.08537        0.33871
#> 9          1        0       0 0.00000    0.0     0.0    1         0.5 0.09756        0.06452
#> 10         1        0       0 0.30435    0.2     0.0    1         0.0 0.10976        0.29032
#> 11         1        0       0 0.34783    0.2     0.5    1         0.0 0.12195        0.30645
#> 12         1        0       0 0.69565    0.6     1.0    1         0.5 0.13415        0.58065
#> 13         1        0       0 0.43478    0.0     0.5    1         0.0 0.14634        0.12903
#> 14         1        0       0 0.34783    0.4     1.0    1         0.0 0.15854        0.45161
#> 15         1        0       0 0.43478    0.2     0.5    1         0.5 0.17073        0.32258
#> 16         1        0       0 0.26087    0.2     1.0    1         0.0 0.18293        0.27419
#> 17         1        0       0 0.39130    0.8     1.0    1         0.5 0.19512        0.59677
#> 18         1        0       0 0.21739    0.0     0.0    1         0.5 0.20732        0.09677
#> 19         1        0       0 0.17391    0.4     1.0    1         0.5 0.21951        0.43548
#> 20         1        0       0 1.00000    0.0     0.0    1         0.5 0.23171        0.25806
#> 21         1        0       0 0.82609    0.0     0.0    1         0.5 0.24390        0.20968
#> 22         1        0       0 0.60870    0.2     1.0    1         0.0 0.25610        0.37097
#> 23         1        0       0 0.30435    0.2     0.0    1         1.0 0.26829        0.29032
#> 24         1        0       0 0.65217    0.0     0.0    1         0.5 0.28049        0.17742
#> 25         1        0       0 0.26087    0.2     0.5    1         0.5 0.29268        0.27419
#> 26         1        0       0 0.82609    0.2     0.0    1         1.0 0.30488        0.41935
#> 27         1        0       0 0.73913    0.2     0.0    1         1.0 0.31707        0.40323
#> 28         1        0       0 0.56522    0.4     0.0    1         1.0 0.32927        0.48387
#> 29         1        0       0 0.30435    0.6     0.5    1         1.0 0.34146        0.53226
#> 30         1        0       0 0.39130    0.6     1.0    1         0.0 0.35366        0.54839
#> 31         1        0       0 0.47826    0.0     0.0    1         0.5 0.36585        0.14516
#> 32         1        0       0 0.56522    0.2     0.5    1         0.0 0.37805        0.35484
#> 33         1        0       0 0.91304    0.0     0.5    1         0.0 0.39024        0.24194
#> 34         1        0       0 0.47826    0.2     0.0    1         1.0 0.40244        0.33871
#> 35         1        0       0 0.78261    0.0     0.5    1         0.0 0.41463        0.19355
#> 36         1        0       0 0.60870    0.2     0.0    1         1.0 0.42683        0.37097
#> 37         1        0       0 0.65217    0.0     0.0    1         0.5 0.43902        0.17742
#> 38         1        0       0 0.56522    0.4     0.5    1         1.0 0.45122        0.48387
#> 39         1        0       0 0.39130    0.4     0.0    1         0.0 0.46341        0.46774
#> 40         1        0       0 0.30435    0.0     0.0    1         0.5 0.47561        0.11290
#> 41         1        0       0 0.78261    0.4     0.0    1         1.0 0.48780        0.51613
#> 42         1        0       0 0.60870    0.0     0.0    1         0.0 0.50000        0.16129
#> 43         1        0       0 0.86957    0.0     0.0    1         0.0 0.51220        0.22581
#> 44         1        0       0 0.69565    0.2     0.5    1         0.5 0.52439        0.38710
#> 45         0        1       0 0.39130    0.0     0.0    1         0.0 0.53659        0.69355
#> 46         0        1       0 0.69565    0.0     0.5    1         0.0 0.54878        0.75806
#> 47         0        1       0 0.30435    0.2     0.0    1         1.0 0.56098        0.80645
#> 48         0        1       0 0.26087    0.6     1.0    1         0.0 0.57317        0.96774
#> 49         0        1       0 0.21739    0.2     1.0    1         0.0 0.58537        0.77419
#> 50         0        1       0 0.73913    0.4     0.0    1         1.0 0.59756        0.95161
#> 51         0        1       0 0.13043    0.4     0.5    1         1.0 0.60976        0.88710
#> 52         0        1       0 0.65217    0.8     0.5    1         1.0 0.62195        0.98387
#> 53         0        1       0 0.26087    0.4     0.5    1         0.5 0.63415        0.90323
#> 54         0        1       0 0.30435    0.0     0.0    1         0.5 0.64634        0.66129
#> 55         0        1       0 0.34783    0.2     0.0    1         1.0 0.65854        0.82258
#> 56         0        1       0 0.65217    0.2     0.0    1         1.0 0.67073        0.87097
#> 57         0        1       0 0.30435    0.2     0.5    1         0.0 0.68293        0.80645
#> 58         0        1       0 0.30435    0.2     0.0    1         1.0 0.69512        0.80645
#> 59         0        1       0 0.30435    0.0     0.0    1         0.5 0.70732        0.66129
#> 60         0        1       0 0.26087    0.2     0.0    1         1.0 0.71951        0.79032
#> 61         0        1       0 0.60870    0.2     0.0    1         1.0 0.73171        0.85484
#> 62         0        1       0 0.17391    0.0     0.0    1         0.5 0.74390        0.64516
#> 63         0        1       0 0.56522    0.0     0.0    1         0.5 0.75610        0.74194
#> 64         0        1       0 0.43478    0.2     0.0    1         1.0 0.76829        0.83871
#> 65         0        1       0 0.21739    0.2     0.5    1         0.0 0.78049        0.77419
#> 66         0        1       0 0.47826    0.0     0.0    1         0.5 0.79268        0.72581
#> 67         0        1       0 0.00000    0.0     0.0    1         0.5 0.80488        0.61290
#> 68         0        1       0 0.30435    0.4     0.5    1         1.0 0.81707        0.91935
#> 69         0        1       0 0.69565    0.4     0.0    1         1.0 0.82927        0.93548
#> 70         0        1       0 0.17391    0.0     0.5    1         0.0 0.84146        0.64516
#> 71         0        1       0 0.47826    0.0     0.5    1         0.0 0.85366        0.72581
#> 72         0        1       0 0.17391    0.0     0.0    1         0.5 0.86585        0.64516
#> 73         0        1       0 0.43478    0.0     0.0    1         0.5 0.87805        0.70968
#> 74         0        1       0 0.73913    1.0     0.0    1         1.0 0.89024        1.00000
#> 75         0        1       0 0.21739    0.2     0.0    1         1.0 0.90244        0.77419
#> 76         0        1       0 0.43478    0.0     0.0    1         0.5 0.91463        0.70968
#> 77         0        1       0 0.43478    0.2     0.0    1         0.5 0.92683        0.83871
#> 78         0        1       0 0.17391    0.0     0.5    1         0.0 0.93902        0.64516
#> 79         0        1       0 0.43478    0.0     0.0    1         0.5 0.95122        0.70968
#> 80         0        1       0 0.56522    0.0     0.0    1         0.5 0.96341        0.74194
#> 81         0        1       0 0.60870    0.2     1.0    1         0.0 0.97561        0.85484
#> 82         0        1       0 0.34783    0.0     0.0    1         0.5 0.98780        0.67742
#> 83         0        1       0 0.08696    0.0     0.0    1         0.5 1.00000        0.62903
#> 84         0        0       1 0.21739    1.0     1.0    0         0.0 0.00000        0.03226
#> 85         0        0       1 0.91304    0.0     0.0    0         0.0 0.01220        0.00000
#> 86         0        0       1 0.78261    1.0     1.0    0         0.0 0.02439        0.04839
#> 87         0        0       1 0.56522    0.6     0.0    0         0.5 0.03659        0.01613
#> 88         1        0       0 0.60870    0.4     1.0    0         0.0 0.04878        0.50000
#> 89         1        0       0 0.65217    0.6     0.5    0         0.5 0.06098        0.56452
#> 90         1        0       0 0.08696    0.0     0.0    0         0.0 0.07317        0.08065
#> 91         1        0       0 0.47826    0.2     1.0    0         0.0 0.08537        0.33871
#> 92         1        0       0 0.00000    0.0     0.0    0         0.5 0.09756        0.06452
#> 93         1        0       0 0.30435    0.2     0.0    0         0.5 0.10976        0.29032
#> 94         1        0       0 0.34783    0.2     0.0    0         0.0 0.12195        0.30645
#> 95         1        0       0 0.69565    0.6     0.5    0         0.5 0.13415        0.58065
#> 96         1        0       0 0.43478    0.0     0.0    0         0.0 0.14634        0.12903
#> 97         1        0       0 0.34783    0.4     0.0    0         0.5 0.15854        0.45161
#> 98         1        0       0 0.43478    0.2     0.5    0         0.0 0.17073        0.32258
#> 99         1        0       0 0.26087    0.2     0.5    0         0.0 0.18293        0.27419
#> 100        1        0       0 0.39130    0.8     0.0    0         1.0 0.19512        0.59677
#> 101        1        0       0 0.21739    0.0     0.0    0         0.0 0.20732        0.09677
#> 102        1        0       0 0.17391    0.4     0.0    0         0.5 0.21951        0.43548
#> 103        1        0       0 1.00000    0.0     0.0    0         0.0 0.23171        0.25806
#> 104        1        0       0 0.82609    0.0     0.0    0         0.0 0.24390        0.20968
#> 105        1        0       0 0.60870    0.2     0.0    0         0.0 0.25610        0.37097
#> 106        1        0       0 0.30435    0.2     0.0    0         0.0 0.26829        0.29032
#> 107        1        0       0 0.65217    0.0     0.0    0         0.0 0.28049        0.17742
#> 108        1        0       0 0.26087    0.2     0.0    0         0.5 0.29268        0.27419
#> 109        1        0       0 0.82609    0.2     0.0    0         0.0 0.30488        0.41935
#> 110        1        0       0 0.73913    0.2     0.0    0         0.0 0.31707        0.40323
#> 111        1        0       0 0.56522    0.4     0.0    0         0.0 0.32927        0.48387
#> 112        1        0       0 0.30435    0.6     0.0    0         1.0 0.34146        0.53226
#> 113        1        0       0 0.39130    0.6     0.5    0         0.5 0.35366        0.54839
#> 114        1        0       0 0.47826    0.0     0.0    0         0.0 0.36585        0.14516
#> 115        1        0       0 0.56522    0.2     0.5    0         0.0 0.37805        0.35484
#> 116        1        0       0 0.91304    0.0     0.5    0         0.0 0.39024        0.24194
#> 117        1        0       0 0.47826    0.2     0.0    0         0.0 0.40244        0.33871
#> 118        1        0       0 0.78261    0.0     0.0    0         0.0 0.41463        0.19355
#> 119        1        0       0 0.60870    0.2     0.0    0         0.0 0.42683        0.37097
#> 120        1        0       0 0.65217    0.0     0.0    0         0.0 0.43902        0.17742
#> 121        1        0       0 0.56522    0.4     1.0    0         0.0 0.45122        0.48387
#> 122        1        0       0 0.39130    0.4     0.0    0         1.0 0.46341        0.46774
#> 123        1        0       0 0.30435    0.0     0.5    0         0.0 0.47561        0.11290
#> 124        1        0       0 0.78261    0.4     0.5    0         0.0 0.48780        0.51613
#> 125        1        0       0 0.60870    0.0     0.0    0         0.0 0.50000        0.16129
#> 126        1        0       0 0.86957    0.0     0.0    0         0.0 0.51220        0.22581
#> 127        1        0       0 0.69565    0.2     0.0    0         0.0 0.52439        0.38710
#> 128        0        1       0 0.39130    0.0     0.5    0         0.0 0.53659        0.69355
#> 129        0        1       0 0.69565    0.0     0.0    0         0.0 0.54878        0.75806
#> 130        0        1       0 0.30435    0.2     0.5    0         0.0 0.56098        0.80645
#> 131        0        1       0 0.26087    0.6     1.0    0         0.5 0.57317        0.96774
#> 132        0        1       0 0.21739    0.2     0.5    0         0.0 0.58537        0.77419
#> 133        0        1       0 0.73913    0.4     0.5    0         0.0 0.59756        0.95161
#> 134        0        1       0 0.13043    0.4     1.0    0         0.5 0.60976        0.88710
#> 135        0        1       0 0.65217    0.8     0.5    0         0.5 0.62195        0.98387
#> 136        0        1       0 0.26087    0.4     0.5    0         0.5 0.63415        0.90323
#> 137        0        1       0 0.30435    0.0     0.5    0         0.0 0.64634        0.66129
#> 138        0        1       0 0.34783    0.2     0.5    0         0.0 0.65854        0.82258
#> 139        0        1       0 0.65217    0.2     0.5    0         0.0 0.67073        0.87097
#> 140        0        1       0 0.30435    0.2     0.5    0         0.5 0.68293        0.80645
#> 141        0        1       0 0.30435    0.2     1.0    0         0.0 0.69512        0.80645
#> 142        0        1       0 0.30435    0.0     0.5    0         0.0 0.70732        0.66129
#> 143        0        1       0 0.26087    0.2     0.5    0         0.0 0.71951        0.79032
#> 144        0        1       0 0.60870    0.2     1.0    0         0.0 0.73171        0.85484
#> 145        0        1       0 0.17391    0.0     0.5    0         0.0 0.74390        0.64516
#> 146        0        1       0 0.56522    0.0     0.0    0         0.0 0.75610        0.74194
#> 147        0        1       0 0.43478    0.2     0.0    0         0.0 0.76829        0.83871
#> 148        0        1       0 0.21739    0.2     0.0    0         0.5 0.78049        0.77419
#> 149        0        1       0 0.47826    0.0     0.0    0         0.0 0.79268        0.72581
#> 150        0        1       0 0.00000    0.0     0.0    0         0.5 0.80488        0.61290
#> 151        0        1       0 0.30435    0.4     1.0    0         0.0 0.81707        0.91935
#> 152        0        1       0 0.69565    0.4     0.5    0         0.5 0.82927        0.93548
#> 153        0        1       0 0.17391    0.0     0.0    0         0.0 0.84146        0.64516
#> 154        0        1       0 0.47826    0.0     0.5    0         0.0 0.85366        0.72581
#> 155        0        1       0 0.17391    0.0     0.0    0         0.0 0.86585        0.64516
#> 156        0        1       0 0.43478    0.0     0.0    0         0.5 0.87805        0.70968
#> 157        0        1       0 0.21739    0.2     0.0    0         1.0 0.90244        0.77419
#> 158        0        1       0 0.43478    0.0     0.0    0         0.0 0.91463        0.70968
#> 159        0        1       0 0.43478    0.2     1.0    0         0.0 0.92683        0.83871
#> 160        0        1       0 0.17391    0.0     0.0    0         0.0 0.93902        0.64516
#> 161        0        1       0 0.43478    0.0     0.0    0         0.0 0.95122        0.70968
#> 162        0        1       0 0.56522    0.0     0.0    0         0.0 0.96341        0.74194
#> 163        0        1       0 0.60870    0.2     0.0    0         0.0 0.97561        0.85484
#> 164        0        1       0 0.34783    0.0     0.0    0         0.5 0.98780        0.67742
#> 165        0        1       0 0.08696    0.0     0.0    0         0.5 1.00000        0.62903
#> 166        0        0       1 0.21739    1.0     1.0    0         0.0 0.00000        0.03226
#> 167        0        0       1 0.91304    0.0     0.0    0         0.0 0.01220        0.00000
#> 168        0        0       1 0.78261    1.0     1.0    0         0.0 0.02439        0.04839
#> 169        0        0       1 0.56522    0.6     0.0    0         1.0 0.03659        0.01613
#> 170        1        0       0 0.60870    0.4     0.0    0         0.0 0.04878        0.50000
#> 171        1        0       0 0.65217    0.6     0.0    0         1.0 0.06098        0.56452
#> 172        1        0       0 0.08696    0.0     0.0    0         0.0 0.07317        0.08065
#> 173        1        0       0 0.47826    0.2     0.0    0         0.5 0.08537        0.33871
#> 174        1        0       0 0.00000    0.0     0.5    0         0.0 0.09756        0.06452
#> 175        1        0       0 0.30435    0.2     0.0    0         0.5 0.10976        0.29032
#> 176        1        0       0 0.34783    0.2     0.0    0         0.5 0.12195        0.30645
#> 177        1        0       0 0.69565    0.6     0.0    0         0.5 0.13415        0.58065
#> 178        1        0       0 0.43478    0.0     0.0    0         0.0 0.14634        0.12903
#> 179        1        0       0 0.34783    0.4     0.0    0         1.0 0.15854        0.45161
#> 180        1        0       0 0.43478    0.2     0.5    0         0.0 0.17073        0.32258
#> 181        1        0       0 0.26087    0.2     0.0    0         0.0 0.18293        0.27419
#> 182        1        0       0 0.39130    0.8     0.5    0         1.0 0.19512        0.59677
#> 183        1        0       0 0.21739    0.0     0.5    0         0.0 0.20732        0.09677
#> 184        1        0       0 0.17391    0.4     0.5    0         0.5 0.21951        0.43548
#> 185        1        0       0 1.00000    0.0     0.5    0         0.0 0.23171        0.25806
#> 186        1        0       0 0.82609    0.0     0.0    0         0.0 0.24390        0.20968
#> 187        1        0       0 0.60870    0.2     0.0    0         0.0 0.25610        0.37097
#> 188        1        0       0 0.30435    0.2     1.0    0         0.0 0.26829        0.29032
#> 189        1        0       0 0.65217    0.0     0.0    0         0.5 0.28049        0.17742
#> 190        1        0       0 0.26087    0.2     0.0    0         1.0 0.29268        0.27419
#> 191        1        0       0 0.82609    0.2     0.0    0         0.0 0.30488        0.41935
#> 192        1        0       0 0.73913    0.2     0.0    0         0.0 0.31707        0.40323
#> 193        1        0       0 0.56522    0.4     0.0    0         0.0 0.32927        0.48387
#> 194        1        0       0 0.30435    0.6     1.0    0         0.5 0.34146        0.53226
#> 195        1        0       0 0.39130    0.6     0.5    0         0.5 0.35366        0.54839
#> 196        1        0       0 0.47826    0.0     0.0    0         0.0 0.36585        0.14516
#> 197        1        0       0 0.56522    0.2     0.0    0         0.0 0.37805        0.35484
#> 198        1        0       0 0.91304    0.0     0.0    0         0.0 0.39024        0.24194
#> 199        1        0       0 0.47826    0.2     1.0    0         0.0 0.40244        0.33871
#> 200        1        0       0 0.78261    0.0     0.0    0         0.0 0.41463        0.19355
#> 201        1        0       0 0.60870    0.2     0.0    0         0.0 0.42683        0.37097
#> 202        1        0       0 0.65217    0.0     0.0    0         0.0 0.43902        0.17742
#> 203        1        0       0 0.56522    0.4     1.0    0         0.0 0.45122        0.48387
#> 204        1        0       0 0.39130    0.4     0.0    0         0.5 0.46341        0.46774
#> 205        1        0       0 0.30435    0.0     0.0    0         0.0 0.47561        0.11290
#> 206        1        0       0 0.78261    0.4     0.0    0         0.0 0.48780        0.51613
#> 207        1        0       0 0.60870    0.0     0.0    0         0.0 0.50000        0.16129
#> 208        1        0       0 0.86957    0.0     0.0    0         0.0 0.51220        0.22581
#> 209        1        0       0 0.69565    0.2     0.0    0         0.0 0.52439        0.38710
#> 210        0        1       0 0.39130    0.0     0.0    0         0.0 0.53659        0.69355
#> 211        0        1       0 0.69565    0.0     0.0    0         0.5 0.54878        0.75806
#> 212        0        1       0 0.30435    0.2     0.5    0         0.0 0.56098        0.80645
#> 213        0        1       0 0.26087    0.6     1.0    0         0.0 0.57317        0.96774
#> 214        0        1       0 0.21739    0.2     0.5    0         0.0 0.58537        0.77419
#> 215        0        1       0 0.73913    0.4     0.5    0         0.0 0.59756        0.95161
#> 216        0        1       0 0.13043    0.4     1.0    0         0.0 0.60976        0.88710
#> 217        0        1       0 0.65217    0.8     1.0    0         0.5 0.62195        0.98387
#> 218        0        1       0 0.26087    0.4     1.0    0         0.0 0.63415        0.90323
#> 219        0        1       0 0.30435    0.0     0.0    0         0.5 0.64634        0.66129
#> 220        0        1       0 0.34783    0.2     0.5    0         0.5 0.65854        0.82258
#> 221        0        1       0 0.65217    0.2     0.0    0         0.5 0.67073        0.87097
#> 222        0        1       0 0.30435    0.2     1.0    0         0.0 0.68293        0.80645
#> 223        0        1       0 0.30435    0.2     0.5    0         0.0 0.69512        0.80645
#> 224        0        1       0 0.30435    0.0     0.0    0         0.0 0.70732        0.66129
#> 225        0        1       0 0.26087    0.2     0.5    0         0.0 0.71951        0.79032
#> 226        0        1       0 0.60870    0.2     0.5    0         0.0 0.73171        0.85484
#> 227        0        1       0 0.17391    0.0     0.5    0         0.0 0.74390        0.64516
#> 228        0        1       0 0.56522    0.0     0.0    0         0.0 0.75610        0.74194
#> 229        0        1       0 0.43478    0.2     0.5    0         0.0 0.76829        0.83871
#> 230        0        1       0 0.21739    0.2     0.0    0         1.0 0.78049        0.77419
#> 231        0        1       0 0.47826    0.0     0.5    0         0.0 0.79268        0.72581
#> 232        0        1       0 0.00000    0.0     0.0    0         0.0 0.80488        0.61290
#> 233        0        1       0 0.30435    0.4     1.0    0         0.0 0.81707        0.91935
#> 234        0        1       0 0.69565    0.4     0.0    0         1.0 0.82927        0.93548
#> 235        0        1       0 0.17391    0.0     0.5    0         0.0 0.84146        0.64516
#> 236        0        1       0 0.47826    0.0     0.0    0         0.0 0.85366        0.72581
#> 237        0        1       0 0.17391    0.0     0.5    0         0.0 0.86585        0.64516
#> 238        0        1       0 0.43478    0.0     0.0    0         0.0 0.87805        0.70968
#> 239        0        1       0 0.73913    1.0     0.0    0         1.0 0.89024        1.00000
#> 240        0        1       0 0.21739    0.2     0.5    0         0.5 0.90244        0.77419
#> 241        0        1       0 0.43478    0.0     0.5    0         0.0 0.91463        0.70968
#> 242        0        1       0 0.43478    0.2     0.0    0         0.5 0.92683        0.83871
#> 243        0        1       0 0.17391    0.0     0.0    0         0.5 0.93902        0.64516
#> 244        0        1       0 0.43478    0.0     0.0    0         0.5 0.95122        0.70968
#> 245        0        1       0 0.56522    0.0     0.0    0         0.0 0.96341        0.74194
#> 246        0        1       0 0.60870    0.2     1.0    0         0.0 0.97561        0.85484
#> 247        0        1       0 0.34783    0.0     0.0    0         0.5 0.98780        0.67742
#> 248        0        1       0 0.08696    0.0     0.0    0         0.5 1.00000        0.62903

# Example with a custom dataframe
df<-data.frame(numeric_var=c(1,2,3,4,5),
               factor_var=factor(c('A','B','A','B','C')))
recode_scale_dummy(df)
#>   A B C
#> 1 1 0 0
#> 2 0 1 0
#> 3 1 0 0
#> 4 0 1 0
#> 5 0 0 1