Replace outliers with NA using IQR fences

Replaces values outside the boxplot fences with NA. The fences are computed as \(Q1 - 1.5 \times IQR\) and \(Q3 + 1.5 \times IQR\), where \(Q1\) and \(Q3\) are the quantiles specified by probs. Designed to be applied across columns with sapply.

Usage

remove_outliers(vector, probs = c(0.25, 0.75), na.rm = TRUE, ...)

Arguments

vector

Numeric vector.

probs

Length-2 numeric vector giving the lower and upper quantile probabilities used to define the fence boundaries. Default is c(0.25, 0.75) (standard quartiles).

na.rm

Logical; whether to remove NA values when computing quantiles and IQR. Default is TRUE.

...

Additional arguments passed to quantile.

Value

A numeric vector the same length as vector with outlying values replaced by NA.

Examples

vector<-generate_missing(rnorm(1000))
df<-generate_missing(mtcars[,1:2])
remove_outliers(vector)
#>    [1]  1.7638487  0.5513212 -1.4988266 -0.1171545  0.0315898 -0.2236617 -0.8701860 -0.4313549  0.3890758  0.8564028  0.8795618  1.9787286  1.8578671 -1.0715946  1.6836451 -0.4233189 -0.5567133
#>   [18]  1.4465413  0.6826160 -1.4819913  1.6614530 -0.6398002  0.4140435 -0.0272033 -0.7966451 -0.9061139  2.0302863  0.2133482 -0.7796140 -1.8006671 -0.2217722  0.3334376 -0.6275975  0.1053362
#>   [35]  1.7753919 -0.9170226  1.5960372  0.3779699 -0.7380740 -1.4018422         NA -0.5058528  0.1693570 -0.0664614 -1.1112059 -0.9150687  0.1570150 -0.6641476  0.5147804 -1.3646466  1.5607374
#>   [52] -1.0005046  1.4586969  0.0755106         NA  0.4264258 -1.0584131  2.0717810  1.0055796 -0.9353865  0.6677904 -0.6096671 -0.7002389 -0.6051017 -1.4150201 -1.4356888  2.3380443  1.5565477
#>   [69]  0.3753320 -1.3393919 -0.5879626  0.8673977 -0.4780236 -1.1791185  0.1887062 -0.0229271  1.5097886 -0.9255308 -0.1753054  0.4959311  0.0944921 -0.3040352  0.3905428 -0.7756180 -0.6711801
#>   [86]  1.7378784         NA  1.2432347 -0.9102313  0.1095978 -1.0094251  0.1955131  0.2760153  0.7526685         NA -1.7843976 -0.4025828 -0.9879521  0.3202164  0.7581798 -0.0993130  0.4755133
#>  [103]  1.7360440  1.0179215 -1.2934187 -0.3282430 -0.9322527  0.3558900 -0.7244783  0.6747390 -0.0006387 -0.4535196 -0.2177017 -0.5580510 -0.5758974 -0.7093144  0.0870887  1.6819700 -0.2213271
#>  [120] -0.4128566  0.1284030 -0.5681684  1.7574751 -0.7158028  1.3189634 -0.1931860 -0.4056225 -0.1928810 -0.2505367 -0.3905384 -1.0851687  1.1345529  0.0359881  1.7835235 -2.3157509  0.9377236
#>  [137] -0.6740845 -1.1886071  1.3964147 -1.1997134 -0.1412586 -0.9517801 -1.2189307  2.0555446 -1.8011344 -1.9807704 -1.7791680  0.2097180 -0.0845167 -0.7995870 -0.4972343 -1.6312612 -0.5390120
#>  [154]  0.5438497  0.9814537  0.3051655 -0.2131227  0.6200917 -0.9456983  0.7736344 -1.6904854  0.3682740  0.1616675  0.7887069  0.2708475 -0.8091630 -0.8593170 -1.6369167  0.1443175 -0.2220557
#>  [171] -0.9858699 -0.7601929  0.7481630 -0.5647992 -0.1174951  0.3281305  0.3913180 -0.8690143 -0.0457445 -0.7586451  0.4086208  1.7823369  0.7821573  1.1262606 -0.0402537 -0.4015410  0.1241716
#>  [188] -0.1861932  1.4300096 -1.1929399 -0.1814763  0.1371586 -0.3679408 -0.3695270 -0.0805755 -0.6510882 -1.5919434 -1.1337548  1.3138982 -0.1752360  0.7064699  1.1229381  1.3796966 -0.6090140
#>  [205]  0.7524193 -1.2270556  1.2761118  0.8101185         NA  0.8632766  0.5853389  0.1062371  0.4103243  1.5909744  0.2333199  0.0657359 -1.5813192  0.2495610 -0.6655288 -0.9595968 -1.1599787
#>  [222] -1.5516663  0.7654666 -0.3769114  0.1902462 -0.8149197  1.0220305 -0.2824967  1.7900356 -0.2814793 -0.5378709  0.4757999  0.9705124 -0.0448537  1.4156840 -1.0134833 -0.7506030 -0.4424033
#>  [239]  2.3553561 -0.4989013  1.1733656 -0.0148793  0.9074796 -0.8615997 -0.5283688 -0.2584037  0.0962348 -1.0266356  0.6628270 -0.7719016 -0.9149207 -2.0681504 -0.3386004  1.3490793  0.7391579
#>  [256]  0.1004763  1.2388467 -0.6700326  0.3543639 -1.1334766  0.3067827  0.6347066 -0.1394027 -0.3862666 -0.3490057  1.8631403 -1.1481103 -0.8629749 -1.0270845 -1.1392957 -0.1784928  1.3271696
#>  [273] -1.5585638  0.2554075 -0.7018752 -0.7444849  1.2756732  0.6274143 -0.7682498 -0.8075773 -0.3884689  0.5374848 -0.6100998  1.2096285  0.2949112  1.3400292 -0.6012644  0.0436734 -0.7511344
#>  [290] -0.6961609  0.8475147 -0.9179348 -0.6487073  0.9990488  1.2597949  1.1508032 -1.1439471  0.4784346  0.7111812  0.3905411  0.4176084 -0.7231765 -0.7753130  0.0943096  0.8948397 -0.1069630
#>  [307]  0.5510933  0.5966401  0.6157828 -0.2327214  0.6510555 -0.3721238 -1.2979636 -0.2412149  0.2476294  0.4846035  0.2726616 -0.4364327  1.6519091  0.1312071  0.3533405 -0.7358725  0.8517873
#>  [324]  0.4422876  1.1182747 -0.6830123  1.0345179  0.1517752 -0.2281158  1.0193415  0.0181906  0.4379505  0.1043132 -1.5706857  1.3473614  1.9679131         NA -0.6215835  1.1676994 -1.8600602
#>  [341] -0.2017065 -0.5550530 -1.5354954  0.7151255 -0.0171257 -0.4356438 -1.6051238  0.5963731 -0.5082423 -1.9540924 -0.3696658 -0.0792629 -0.9079000 -1.7877425 -0.0336012 -0.4131472 -0.3521139
#>  [358]  0.1840363 -0.5202087  1.3212294 -1.9806860  0.9212335  0.2416187  1.7428096 -0.0488657 -1.2708158  0.2570999 -2.3369111  0.5737221 -0.0850106 -1.3039075  1.0709715  0.5894103  0.8600716
#>  [375] -0.4358736  1.4240011  0.2570552  0.3951142  0.4423983 -1.0291598  1.3521039 -0.1356927 -1.0827652 -0.4827612  0.5733161 -1.3888176  1.7179440 -0.6123736 -1.4141327  0.2599326  0.7189705
#>  [392]  0.8646054 -0.6831815  0.1693647  0.6064351 -1.9943632  1.3131999  0.3540647  0.4938685 -0.4894048  2.0366110 -0.4288175  0.6982347 -0.2571123 -0.6441409 -0.4172234  0.7514470  0.3247956
#>  [409]  1.2229664  1.0482670 -0.6644836  0.5523544 -0.6850020 -0.2478495  0.6942857 -0.7178061  0.2505752  0.7826933  0.6468171 -0.8078137 -0.2868143 -0.2295448  0.2474643  0.4121058 -0.0561458
#>  [426]  0.3418227  0.3240168  1.4545267  1.1529944 -0.3327361 -0.2156011  0.5867904  0.6427477  0.8306627 -0.4938819 -1.2492334  1.2042656 -0.0641542  2.2037103 -0.1355586 -0.5568683 -0.6651418
#>  [443]  1.6980253 -0.9543132  0.3914136 -1.7868152  1.5069649  2.0488792 -0.6969202 -0.8765905 -2.3200325  0.2316552 -0.0205850 -1.8068636 -1.0353216  1.9350585 -0.5004811 -0.8572144 -0.4748733
#>  [460]  0.9849660 -0.2728316 -1.1416392 -0.2241815 -0.4900848  1.4053805  0.7274225 -1.4430844 -1.9177252 -1.6299641  0.9443752 -0.6214562  1.1147541  0.9259431 -0.6079676         NA         NA
#>  [477]  1.2061509  0.2247215  0.9455002 -0.6584437 -0.7124701 -1.8627612 -1.0378638  0.3206199  1.6228781  0.0141192 -0.1394568 -1.8044401 -0.2535080 -0.1304673 -0.3625088  0.2797290  0.0364198
#>  [494]  1.8244083  1.1745120 -0.7881665 -1.8923681 -0.6444122 -0.7556993 -1.6342306 -0.1216935 -0.3291946  0.7721821  0.9272185 -2.1579721  1.0773120 -1.9234629  0.0204989 -0.2468898 -0.2910075
#>  [511] -0.3495414  1.3999744  1.2065268 -0.9548351  1.0522450 -1.6359168  0.1396332 -0.4217730  1.1525487  1.5781696  0.3547477 -0.2337731 -0.6174432 -0.1139177 -1.9165556 -1.2927942  0.6775238
#>  [528] -0.0308444 -0.8472164  0.4478684  0.6550068  0.7798073  1.2033298 -0.6951621 -0.0242304  0.1151346  2.0637625  0.1506987 -0.6522298 -0.2139489 -0.1767268  0.0900816 -0.7008603 -0.9644507
#>  [545]  0.3574568 -0.2152024 -0.0944919 -0.2539733  1.6260508 -1.6076334 -0.3293157 -0.5468772  0.7583608 -0.0941592 -0.8632591  0.6716703 -1.7193742 -2.6161593  1.5968504 -0.2197119  0.0017571
#>  [562] -0.9257213  1.9516570 -0.5428054 -1.4625931  1.3916718  1.7083655  0.7305287 -0.5981640         NA -0.0302284  0.4599646  1.3977438 -0.6197520  0.6064862 -0.4382398         NA  2.0407365
#>  [579]  0.0204617  1.5709152  0.8379184 -1.4521137  0.5373400 -0.0166437 -0.0573098 -0.9900148 -1.0482119 -1.3403624  0.4853978  0.3727759  2.5276157 -0.8555935  1.4725585 -0.9680554         NA
#>  [596]  1.0723487 -0.3838185 -0.5701183  0.3688552 -0.0196139 -0.7696466 -0.4445464  0.0292289 -0.5246749 -0.1563998  0.2786958  0.4533523 -0.0721284 -0.1625175 -0.2834746  1.3544294  0.3674563
#>  [613]  0.2627538 -0.4189150 -1.2236791 -0.3477349  0.7725040 -0.7740909  0.6427001  0.5391271  0.7204379  0.2714651 -1.9523925 -0.0846142 -0.9711132  1.9485481 -2.0448975  0.0387978  1.0872954
#>  [630] -0.6790374 -0.5723894  0.8232486  0.3058881 -0.5277736  0.0002125 -0.0782511 -0.2821736  0.7671199  0.7493992 -0.5416423 -1.4033322 -1.7764547 -1.4288663 -0.8037157 -0.8880510  0.5714898
#>  [647]  1.2506660  2.2054240 -0.0795508  0.3697794  1.0539690  0.6439107 -0.5869147  0.1500711 -0.4697220 -1.0295534 -1.4674610 -0.1226827  0.1411561 -1.7237551 -1.2957237 -0.0332780  0.2877086
#>  [664] -0.4692574 -0.9547124  0.2876970 -2.2437066 -1.5724114  0.7388383 -0.0091757  0.6642059 -1.9103177  1.2754954  0.0555574 -1.2084181  0.6974655 -0.1019820 -0.4916122 -0.3511454 -0.7200075
#>  [681] -1.5336706  1.0089969 -0.3454971  1.0630222 -0.9101411  0.2831120         NA -1.4887342 -1.8530681  0.1963867  0.1822868 -0.1602405 -2.2534885  0.0195510 -1.0301052  0.2991672  0.8550369
#>  [698] -0.9666045 -0.7434050 -0.8035741  0.2366791 -0.3009390  0.0190865  0.4398741 -2.0983282  0.3271145 -1.0031138 -0.1299974 -0.0691673 -1.0195200  0.1493203 -0.4287710 -0.3582979 -1.8544350
#>  [715]  0.1648345  0.7958558 -0.5898711  0.2455184  0.9496082  1.9968399 -0.3047131  1.6546520  0.0094635  0.2039007 -0.8934715  1.7098061  0.0686904 -0.4940104 -0.6931165  0.8745489 -0.9572686
#>  [732]  0.8133415 -1.9286287  0.1319931 -0.1254901  0.1377756 -0.1682064  1.0960445 -0.5741880  0.5257651 -0.9083381 -0.3228743  0.7003427 -0.4061728 -0.4029317 -0.7458944  0.6775909 -0.4261069
#>  [749] -0.3508113  0.1006969  0.0945062  0.4221432  2.1769477  0.6328749 -1.2839716 -1.7306263  2.3943866  0.1351339  0.1106156 -1.1713615  0.6316614  1.9773084  1.7198549  2.4874911  1.0753499
#>  [766]  0.9273781  0.4210222  1.1641682  1.1152585  0.2556105 -0.9283986  0.3003652 -0.1931387  0.7921634 -0.3175672  1.2208966 -0.6644783  0.2517890 -0.6159774         NA -0.6308578 -0.2329518
#>  [783]  0.0849722 -0.9588499 -0.2907947 -0.3290982 -1.1353983 -0.2134184 -0.2560675  0.1326190 -1.8007598  0.6931759  1.1997149  0.2625792 -1.8900321 -0.8186878 -1.7216518  1.3946590 -0.3076632
#>  [800] -0.2770497 -0.0251709 -0.7131805  0.7678791 -0.0540752 -1.2029475 -0.8856175  0.0222032 -1.9960323  0.6897130 -0.3543719 -0.0516713 -0.5678143 -0.9224179  0.3672475  1.3599305  0.9804906
#>  [817]  0.0837819  0.1473890 -0.1755437 -1.4807158 -0.3921824 -0.4157733 -0.1080223  0.1738225 -0.0129567 -0.0967437 -0.6168450  0.1841993  2.3500963 -0.5351716  0.1168795  1.1072883  0.8180749
#>  [834] -0.5605293 -0.9184201 -0.4701070 -0.2017271         NA  0.5955423 -1.4455274         NA  0.2602139  0.4560344 -1.0517308  1.6128172  0.1010673         NA -0.2148164  2.4904952 -0.7995792
#>  [851] -0.9279710  0.3026504 -1.1529183 -0.8675658  0.4623711  0.6268287  0.3448756  0.7657949  0.3234427 -0.9665915  0.2550240 -0.7420750  0.3088046  0.2959091 -0.5718541  1.3313138  0.5138658
#>  [868] -1.1010449  0.3962257  1.3459251 -0.0075619 -1.1716359  0.6448924  0.8373935 -1.6075843  0.2111530 -1.8085489 -1.9244410  0.3265693 -1.1328914         NA  0.3000982 -0.1327747  0.0850357
#>  [885] -1.0827095  2.0078446 -1.6204878  0.1962896 -0.9439907  2.0276281  1.4973422 -0.5458574  1.6784901  0.2904578  0.9107855 -1.6034126  0.0453345 -0.3933775 -1.1834042 -0.5854769  0.1341776
#>  [902]  0.6480408 -0.6433574 -1.5189685 -0.3941926 -0.8493561  0.0995350  1.3650982  0.1423394  2.1211037  0.0376973 -0.2458094  0.3706439  0.2960177  0.3842043  0.2267208         NA -1.2055394
#>  [919] -2.4223015 -0.6374345  0.1926614  0.3108176  1.9903752 -0.1197080  0.5376153 -1.4409905  0.9655291 -0.1413521  0.3106625  0.4366839 -0.6151543 -1.2128312  1.2982307 -0.2327584  0.3345564
#>  [936] -1.0513178 -0.6849663 -1.0786565  0.3347891  0.6845243 -0.6670224 -0.5515665 -1.4583727  0.3655177 -0.0019316  0.9572086 -1.6063830  0.7724072  0.8162792 -0.9183325 -0.2125741 -1.3421208
#>  [953]  0.0875469  1.4828442 -1.1497107  0.2171819 -0.0730178  0.0447354  0.4343609  1.9099166 -0.5591626 -0.0015672  0.3661700  0.9532805  0.7404440  0.6412898 -0.0684181 -0.4056049 -0.1885456
#>  [970]  0.3833659 -0.1133034 -1.1267185 -0.6482538  1.7136283  0.4424603  0.2034329  0.6547258  0.4597070 -1.0119717  0.2911918  0.6061521 -0.3499933 -0.7195912  0.4491030 -0.0008224  0.0022367
#>  [987]  0.6121041 -0.5369825 -0.5028898  0.8265536  1.0969176  0.0062158 -1.5639743  1.4994895  0.4724455 -0.5953678 -1.3103113  0.1223920  0.1032854 -0.0884095
data.frame(sapply(df,remove_outliers))
#>     mpg cyl
#> 1  21.0   6
#> 2    NA   6
#> 3  22.8  NA
#> 4  21.4   6
#> 5    NA   8
#> 6  18.1   6
#> 7  14.3   8
#> 8    NA   4
#> 9  22.8   4
#> 10 19.2   6
#> 11 17.8   6
#> 12   NA   8
#> 13 17.3  NA
#> 14 15.2   8
#> 15 10.4   8
#> 16 10.4   8
#> 17 14.7  NA
#> 18 32.4   4
#> 19 30.4   4
#> 20 33.9   4
#> 21 21.5   4
#> 22 15.5   8
#> 23 15.2   8
#> 24 13.3   8
#> 25 19.2  NA
#> 26 27.3  NA
#> 27 26.0   4
#> 28 30.4   4
#> 29 15.8   8
#> 30 19.7   6
#> 31   NA   8
#> 32 21.4   4