Replace outliers with NA using IQR fences

Replaces values outside the boxplot fences with NA. The fences are computed as \(Q1 - 1.5 \times IQR\) and \(Q3 + 1.5 \times IQR\), where \(Q1\) and \(Q3\) are the quantiles specified by probs. Designed to be applied across columns with sapply.

Usage

remove_outliers(vector, probs = c(0.25, 0.75), na.rm = TRUE, ...)

Arguments

vector

Numeric vector.

probs

Length-2 numeric vector giving the lower and upper quantile probabilities used to define the fence boundaries. Default is c(0.25, 0.75) (standard quartiles).

na.rm

Logical; whether to remove NA values when computing quantiles and IQR. Default is TRUE.

...

Additional arguments passed to quantile.

Value

A numeric vector the same length as vector with outlying values replaced by NA.

Examples

vector <- generate_missing(rnorm(1000), missing = 10)
df <- generate_missing(mtcars[, 1:2], missing = 10)
remove_outliers(vector)
#>    [1]  0.170678 -1.225071  0.387333  0.490668  2.239126  0.892212  0.798933  1.935443  0.394250  0.632521  0.102591  1.010211  1.214678  0.500871 -0.909468 -2.197209  0.438545 -1.029415  0.713078
#>   [20]  1.028269 -0.570633 -0.434009  1.005317  1.203545  1.496183 -0.156598        NA  0.352539 -1.286189 -0.272523  0.156915 -0.881813 -0.121898 -0.855251 -1.005729 -0.913952 -2.078234 -0.140647
#>   [39]  1.299638 -0.029369 -2.236622  0.545030  0.830842 -0.497579 -0.601027 -0.512198 -0.004932 -0.097148 -0.024035 -0.127390 -0.131886  0.453742 -1.406336  0.210931  0.752336  0.138380  0.159680
#>   [58]  0.228435 -1.162784  0.854659 -0.242096 -2.015995 -0.227690  0.224170 -0.736825 -0.846711 -0.111459 -0.311602 -1.420324 -2.466939  0.484716 -0.937972        NA -0.162945  0.220456  0.876211
#>   [77] -2.148963 -0.233751 -1.998493 -1.234780  0.846304 -0.046495  0.403643  0.767455  0.867102 -1.254581 -0.157798 -0.512317  0.413021  0.564256  0.726066 -0.415102  0.041615  0.224292  0.786409
#>   [96] -0.423363 -0.367534 -1.767020 -0.183538  0.834872 -0.587201 -0.358932  2.114129  0.645411 -0.312320  1.185448  0.553799 -0.053754  2.127463 -0.187706  0.098758  1.910378  1.621456  2.093068
#>  [115] -0.692236 -2.287198 -1.655602  0.569509 -0.471223 -0.033210  0.377367  1.157718  1.398650  2.118504 -1.087460  1.216033  0.677225 -0.726377  0.112844  0.065214 -0.630627  0.518855 -0.802944
#>  [134] -0.681623 -0.037163 -0.287274 -1.026027  1.398359  0.836690 -0.329896  0.534802 -1.207253  0.628767  0.576165  1.236224  0.475540 -1.547147  0.055469  0.680062 -1.350202  0.245102 -0.307349
#>  [153]  0.731365  0.201382 -1.575463  0.023539  0.801641 -2.555627 -0.526911 -1.312859  0.800861  0.713101  1.031915  1.681808 -0.226310 -0.197929  1.043702 -0.441737 -1.636629  0.211563 -0.464832
#>  [172] -0.662357 -0.132954 -1.321702  0.076693 -0.384991  0.947306  0.021686 -0.458294  0.228862 -1.396408  0.924294        NA  0.802891 -0.060525  0.089329  0.401311 -0.888561  1.449169  0.205510
#>  [191] -0.056120 -0.639859  0.717781 -0.064500  0.132430  0.034715  0.281653  0.919278 -1.042723 -0.213623  1.777947  1.460941 -0.364551 -0.613160  0.495907  0.245639 -0.127905 -1.250439        NA
#>  [210]  0.975436  1.084078 -0.359387 -1.015520  1.335521 -0.155562  0.926608  0.889246  0.145117 -0.818647 -1.236823  2.317226  0.050724  0.290836 -1.811290 -1.145148  1.280873 -1.359470  1.884971
#>  [229] -1.368602  0.543940 -1.238030  1.877132 -1.445810 -0.153367 -0.945689  0.312405  0.274724 -0.315442 -0.384368 -1.428936  0.963418  0.613782 -0.368554 -1.807045  1.355438  0.445680  0.164795
#>  [248]  0.029978 -0.676113  1.136050  2.169332 -1.089019 -1.102726  1.302636        NA  0.341312 -0.223550 -0.198772  1.633303 -1.869622 -0.250089 -1.370759  1.451262 -2.137910 -0.577261  0.236449
#>  [267]  1.099857  0.561283  2.303627  2.020896 -0.057879  0.442093 -1.303331 -0.035220  0.924254 -0.360828  0.467173 -0.201237 -0.005347 -0.533003 -0.515415 -0.084911  0.435933  0.997954 -1.163189
#>  [286]  1.680894  1.217121 -0.027439 -0.291546  0.972628 -0.239707 -0.901718 -0.631174 -1.625975 -2.149496  2.112628  0.269080 -0.519472 -0.015734 -0.151253  0.088231  0.778903  0.386720  0.347953
#>  [305] -1.899571 -0.057682 -0.471305  0.634731 -0.960744 -0.804825 -0.495947 -0.418477 -0.654984  1.395486        NA  1.900826 -0.395319 -0.586702  0.461765  0.097209  0.676026 -0.748771  1.025553
#>  [324] -0.696595 -1.277135 -0.092424  1.847464 -0.929629 -1.030800 -1.319514 -0.084206        NA -0.159947  0.863308  0.969371  0.614129  2.496919  1.155265  1.874872 -0.237557 -0.452780 -0.716752
#>  [343]  0.488934  0.834202  0.925865  0.943155 -0.560074  0.687699  0.219825 -1.005237  0.373101  0.469392  0.064626  1.004815 -0.373829  0.936109  0.844989  1.092658  0.629245  1.037235  0.870617
#>  [362] -0.627408  0.435729  0.239407 -1.256120  0.223570  0.504590  2.117015 -0.817492 -0.249266  0.462999  0.667326  0.488170  1.076487  1.149296 -1.622123  1.126283  0.590428 -0.740585  0.252454
#>  [381] -0.336998  0.948040 -0.660350  0.683054  0.078392  0.533368  0.875587  0.138511  0.940188 -0.663677 -0.842412  2.364362 -0.032863  0.650430 -0.698381 -0.437501 -0.719087 -0.274989  0.896918
#>  [400] -0.639743 -0.076132  0.848679  0.617059 -1.988485  0.631595 -1.944145  0.477375  1.778730  0.887893 -1.103164 -0.302634  0.833795 -2.409530  0.257452 -2.173738  0.026168 -0.573890  0.277073
#>  [419]  0.263804  0.889458 -0.531692 -0.288877  2.324238  0.805070 -0.824741  0.614008  0.006184  0.506463 -1.036276 -2.058183 -1.183702 -0.521875 -1.155976  1.707211 -0.401446 -0.078048 -0.689232
#>  [438] -0.703403  0.205670  0.420175  1.063715 -0.727681        NA -1.388424  0.310032  0.190054 -0.163805  0.513491  0.900096  0.119871  0.941527  0.515489 -0.554761  0.772377  0.743596 -0.589652
#>  [457]  0.626300 -1.876366  0.139846  0.240304  0.119982 -2.011711 -0.373660 -0.997170 -1.131102  0.015209  0.403705 -0.541211 -0.784861 -2.560052  0.072801  0.750244 -0.128249 -0.487867  0.908788
#>  [476] -1.852464 -0.310148 -0.313865  0.023551  1.183460 -0.552833  0.597040  0.680969  0.011961 -0.206808  0.738849 -0.025722 -1.796728 -1.815611 -0.257745 -0.060727  0.991874 -0.968441  0.380721
#>  [495]  0.213510 -0.968507  1.494873        NA -1.700628 -1.467678 -0.216349  1.151785  0.633480 -0.277310 -1.743348 -1.018553 -0.309481 -0.514006 -0.741466  0.191217 -1.166984 -1.626141 -0.976897
#>  [514] -0.688621  0.407801 -1.121405 -1.562934  0.234661 -1.983761  0.978032 -0.325947 -0.226813  0.422330 -0.193525 -0.271017 -0.264942 -0.553388 -0.436708  0.325087 -0.782145  0.521032  1.887861
#>  [533]  1.583178 -0.131828 -1.578077 -0.769731 -1.089302 -0.737530  0.902151 -0.349407  0.589930 -0.157808  0.797874  1.086562  0.834957 -0.296866  1.854301  1.298332 -0.535241 -0.884331  1.861033
#>  [552] -0.270592  0.585641 -0.577058 -0.319015 -0.314025 -0.659339 -1.223160  0.109204 -0.234407  0.161945 -0.855904 -1.213610 -0.067721  1.456469 -1.442939 -0.512530  0.243984  1.677512  0.079474
#>  [571] -0.856427 -0.778777 -0.380936 -1.897358 -0.071771 -1.030865  0.670695 -0.361288 -0.279306 -0.645890 -0.515729 -1.374556  1.010378  0.453706  0.274880 -1.271791  0.973595  1.280787 -0.025877
#>  [590] -1.735964 -0.945176 -0.461120  0.559566  0.207372  0.400516 -0.051410  0.824945 -1.678068 -0.923279 -0.657448  1.654942 -1.163929 -0.879143  0.224613  1.420550  0.500921  0.284830 -0.720575
#>  [609]  0.896461 -1.432468  0.385993 -1.110654 -0.075503 -1.387222  0.274714  0.546312 -0.539095 -1.841499 -0.790289 -1.094068  0.463697  0.016765  0.713716  0.372720 -0.631419 -0.345195 -0.233673
#>  [628] -2.402710 -0.841016 -0.525578  0.964676  1.168688  0.037526 -1.059450 -0.203199 -0.260523 -1.503593 -1.191546  2.457074 -1.421405 -1.170671  0.612729  0.504073 -0.033604  1.067788 -1.084364
#>  [647]  0.565756 -0.276091 -0.825709  0.490221 -0.926532  1.602550        NA -0.365513 -0.006332 -2.380531  1.520820 -0.619902 -0.484029 -0.926161 -0.285926  1.587795  0.659089  1.206298 -0.024830
#>  [666] -0.904587 -0.446870  0.125086  1.148091  0.455014  2.021946 -0.695270  1.603965  1.351707  0.439515  1.571622 -1.741253 -1.484015 -0.555233 -1.358605 -0.082657 -0.358236 -1.289118  0.966683
#>  [685] -0.420456  0.778215  0.577273 -0.396722 -1.557470 -0.548408  0.217260  0.913235  1.552215  1.268131  0.760402  0.659180  0.653990  1.555976  0.416611  0.028053 -2.366009  0.546044 -0.087077
#>  [704]  0.368892 -1.440439  0.948342 -1.622469 -1.422041 -2.199560  0.102243 -1.700944  1.667568  0.651328  1.533247 -0.007104  0.064437  1.425220  1.095977 -0.123836 -0.602284 -1.096140 -0.311297
#>  [723]  0.175139  0.530810  0.906411        NA  0.184795 -0.505012  1.090732 -0.185498 -0.471422  0.633122 -0.272124  0.792839  1.051680  0.467591  1.577758  1.192579  0.687809 -0.540566  0.692903
#>  [742] -2.194040 -0.954047 -0.140184  0.367150  0.055858  1.853887  1.233489 -1.533458  0.070456  0.624654 -0.223982  0.351171  0.122204 -1.803557  1.001495  1.199220 -1.426005 -1.154185  0.195879
#>  [761]  0.489033 -0.030151  0.162181  0.106499 -1.176967  0.934429  0.421975 -0.503019 -0.267882  0.937288  0.034024  0.243159  1.143689  0.827406  0.348058 -0.381203  1.361281 -0.885542  0.388984
#>  [780]  0.411905  0.709913 -1.315294 -1.536934 -0.646003 -0.968643 -1.045695 -0.132059  0.138725 -1.609804  0.972991  1.048883 -0.115849  0.910868  2.061831  0.776568 -0.186582 -0.509899  0.714077
#>  [799] -1.035263  1.894849  0.056793 -1.700726 -0.064837  0.893159 -0.237685 -0.883780 -0.268114 -1.425159  0.282729 -0.158980 -0.141044  0.386239 -0.473855  1.181902 -0.846383 -0.231096 -1.541975
#>  [818] -0.125906  0.891539  0.708697  1.365485  0.146941  0.034377  1.350905 -0.364784 -0.820966 -0.447457  1.119657  0.209283  0.583689 -0.686082 -1.018490 -1.134753 -0.458250  0.867376  0.707050
#>  [837]  0.220164  0.317230 -2.308669 -0.764127 -0.257215  1.171747 -0.645359 -1.790562 -1.079786  0.714449 -1.812945  1.048169  0.202716  0.281211  1.032058  2.293879 -0.059256  0.411401 -1.279242
#>  [856] -0.426502  0.849957 -1.803013 -0.969554  1.650679  0.665957  0.361115 -0.487799  1.428382 -0.904253 -2.093694  0.921679 -1.849023 -1.659894 -0.276360 -0.828434 -0.686071 -0.271608  1.243811
#>  [875] -0.172021 -2.546475  1.640627  0.970237  0.020307 -2.651398  1.117812 -0.819083  1.704550  0.849856  0.543337  0.056035  1.584050 -1.438635 -0.639719  0.086765  0.466811  0.417253  0.607895
#>  [894]  0.280210 -1.176268 -1.294234 -0.318741  0.845858  0.059042  0.280053 -0.836591  0.653533 -0.902541 -2.032552  1.342897 -0.099605 -1.320546  0.490032 -0.137010 -0.054302 -0.644716  0.020392
#>  [913] -0.270451 -0.472232 -2.137675  0.097327 -0.379329  0.434193 -0.115664 -1.298970 -1.447406  0.117877 -0.713257 -1.510396 -0.291504 -0.890669 -0.685477 -0.063932 -0.269658        NA  1.051274
#>  [932]  1.291020  1.238249 -1.901656 -0.442676  1.237265 -0.034629  0.870558 -1.358354  0.215011  0.449079  1.204462  1.162177  0.190854 -0.559199 -1.184502 -0.112966  1.029247 -0.687589        NA
#>  [951] -0.147656  0.427034 -1.486970  0.558629 -0.235440  0.215798 -0.071743 -1.651181 -0.733304 -0.001886  0.700393        NA -1.024710 -0.374095 -0.743649 -0.674708 -0.676335 -0.973638 -0.764557
#>  [970]  0.979196  1.173833  1.374666  1.202738  0.026949  0.711260 -0.091452 -0.116862 -1.125789  1.139871 -0.459491 -0.920665 -0.303931  0.267685 -1.256961  1.552993        NA  1.430688  1.402933
#>  [989]  0.020715  1.222090 -0.577104  0.825420  0.382837 -1.211663  1.116297 -0.818012 -0.444865  0.844661  0.951259 -0.862401
data.frame(sapply(df, remove_outliers))
#>     mpg cyl
#> 1    NA   6
#> 2  21.0   6
#> 3    NA  NA
#> 4  21.4  NA
#> 5  18.7  NA
#> 6  18.1  NA
#> 7  14.3   8
#> 8    NA   4
#> 9  22.8   4
#> 10 19.2   6
#> 11 17.8   6
#> 12 16.4   8
#> 13 17.3   8
#> 14 15.2   8
#> 15 10.4  NA
#> 16 10.4  NA
#> 17   NA   8
#> 18   NA   4
#> 19   NA   4
#> 20   NA  NA
#> 21 21.5   4
#> 22 15.5  NA
#> 23   NA   8
#> 24   NA   8
#> 25 19.2   8
#> 26 27.3  NA
#> 27   NA   4
#> 28   NA   4
#> 29 15.8   8
#> 30 19.7  NA
#> 31   NA   8
#> 32   NA   4